Added input and output resource directories. Added javascript relocation config file. Added check on the copy_tree function to avoid copying non-existent directories. Added generate_js_config and rooterize_path functions. Create resource folder on init. Clear resource folder before resolving again. Clear resource folder on clear.

+++ src/utils.py
@@ -25,26 +25,27 @@ def copy_tree(source:str, destination:str, exceptions:list[str|list[str]] = []):
     exceptions:list[list[str]] = list(map(lambda x: parts_in_path(x) if isinstance(x, str) else x, exceptions))
 
     os.makedirs(destination)
-    listed = os.listdir(source)
-    while listed:
-        item = listed.pop(0)
-        item_parts = parts_in_path(item)
+    if os.path.exists(source):
+        listed = os.listdir(source)
+        while listed:
+            item = listed.pop(0)
+            item_parts = parts_in_path(item)
 
-        is_exceptional = False
-        for exception in exceptions:
-            if item_parts[-len(exception):] == exception:
-                is_exceptional = True
-        if is_exceptional:
-            continue
+            is_exceptional = False
+            for exception in exceptions:
+                if item_parts[-len(exception):] == exception:
+                    is_exceptional = True
+            if is_exceptional:
+                continue
 
-        src = os.path.join(source, item)
-        dst = os.path.join(destination, item)
+            src = os.path.join(source, item)
+            dst = os.path.join(destination, item)
 
-        if os.path.isfile(src):
-            copy_file(src, dst)
-        elif os.path.isdir(src):
-            os.mkdir(dst)
-            listed.extend(map(lambda x: os.path.join(item, x), os.listdir(src)))
+            if os.path.isfile(src):
+                copy_file(src, dst)
+            elif os.path.isdir(src):
+                os.mkdir(dst)
+                listed.extend(map(lambda x: os.path.join(item, x), os.listdir(src)))
 
 
 def delete_folder(folder_path:str):
@@ -79,6 +80,13 @@ def generate_php_config(filename:str, definitions:dict[str, str] = {}, sdir:bool
         
         for func in functions:
             php.write(getattr(PHPFunctions, func) + '\n')
+
+def generate_js_config(filename:str, definitions:dict[str, str] = {}):
+    with open(filename, 'w') as js:
+        js.write(f'window.relocation = ' + json.dumps(definitions))
+
+def rooterize_path(path:str):
+    return path if path[0] == '/' else '/' + path
 
 def binary_search(lst:list[Any], searchFor:Any, key = lambda x:x):
     llst = len(lst)
